+++ NEWS.md
@@ -1,4 +1,4 @@
-# version 1.6-28
+# version 1.6-31
 
 ## new